Add persistent cloud db support
* Disable pre-loading the database in production
