+++ src/main/resources/application-production.properties
@@ -1,0 +1,1 @@
+spring.jpa.hibernate.ddl-auto=update
